[BE_refactor] OrderStatus Enum을 Order 엔티티에서 분리
