+++ server/src/main/java/com/example/server/order/entity/Order.java
@@ -1,5 +1,6 @@
 package com.example.server.order.entity;
 
+import com.example.server.order.data.OrderStatus;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -35,21 +36,7 @@ public class Order {
 
   private OrderStatus status;
 
-  public enum OrderStatus {
-    ORDER_COMPLETED("주문완료"),
-    ORDER_CANCELED("주문취소"),
-    DELIVERY_IN_PROGRESS("배송중"),
-    DELIVERY_COMPLETED("배송완료"),
-    REFUND_APPLIED("환불대기중"),
-    REFUNDED("환불완료");
 
-    @Getter
-    private String status;
-
-    OrderStatus(String status) {
-      this.status = status;
-    }
-  }
 
   // User 연관관계 매핑 필요 아직 User 엔티티가 없음
   // OrderMealBoxy 연관관계 매핑 필요
